requirement.txt
Django==1.9.5
django-push-notifications==1.4.1
djangorestframework==3.3.3
mysql-python==1.2.5
six==1.10.0
wheel==0.24.0
